Divide Interceptor in SuccessInterceptor & ErrorInterceptor
